Added Finite State Machine, ContactScenario.
ContactScenarios are objects that are created on a arrow collision event and sent to the GameStateManager. This will be used to ascertain what happens. Buttons will be targets just the same as birds, and will fire off the correct callbacks to be used as menu buttons.
Also, a bow dock controller has been implemented with an automatic callback feature. If you drop the bow, and you wait 2 seconds, it will reset the bow back to the dock and disable it's physics, ready to be picked up again. However, if you pick the bow back up before the timer to redock ends, then it will cancel the redock timer all together as if you didn't drop it.